+++ src/deploy/functions/prepare.ts
@@ -4,6 +4,7 @@ import * as args from "./args";
 import * as proto from "../../gcp/proto";
 import * as backend from "./backend";
 import * as build from "./build";
+import * as experiments from "../../experiments";
 import * as ensureApiEnabled from "../../ensureApiEnabled";
 import * as functionsConfig from "../../functionsConfig";
 import * as functionsEnv from "../../functions/env";
@@ -236,13 +237,17 @@ export async function prepare(
       const exportType = backend.someEndpoint(wantBackend, (e) => e.platform === "run")
         ? "tar.gz"
         : "zip";
+
+      const isDart = supported.runtimeIsLanguage(wantBuilds[codebase].runtime!, "dart");
+      const executablePaths = isDart ? ["bin/server"] : [];
+
       const packagedSource = await prepareFunctionsUpload(
         options.config.projectDir,
         sourceDir,
         localCfg,
         [...schPathSet],
         undefined,
-        { exportType },
+        { exportType, executablePaths },
       );
       source.functionsSourceV2 = packagedSource?.pathToSource;
       source.functionsSourceV2Hash = packagedSource?.hash;
@@ -497,14 +502,17 @@ export async function loadCodebases(
       throw new FirebaseError(
         `Functions codebase ${codebase} has invalid runtime ` +
           `${firebaseJsonRuntime} specified in firebase.json. Valid values are: \n` +
-          Object.keys(supported.RUNTIMES)
+          (Object.keys(supported.RUNTIMES) as supported.Runtime[])
+            .filter((runtime) => !supported.isDecommissioned(runtime))
             .map((s) => `- ${s}`)
             .join("\n"),
       );
     }
     const runtimeDelegate = await runtimes.getRuntimeDelegate(delegateContext);
     logger.debug(`Validating ${runtimeDelegate.language} source`);
-    supported.guardVersionSupport(runtimeDelegate.runtime);
+    if (!experiments.isEnabled("bypassfunctionsdeprecationcheck")) {
+      supported.guardVersionSupport(runtimeDelegate.runtime);
+    }
     await runtimeDelegate.validate();
     logger.debug(`Building ${runtimeDelegate.language} source`);
     await runtimeDelegate.build();