+++ src/init/features/apphosting.ts
@@ -17,7 +17,9 @@ import { FirebaseError } from "../../error";
 import { AppHostingSingle } from "../../firebaseConfig";
 import { ensureApiEnabled } from "../../gcp/apphosting";
 import { isBillingEnabled } from "../../gcp/cloudbilling";
+import { Options } from "../../options";
 import { input, select } from "../../prompt";
+
 import { readTemplateSync } from "../../templates";
 import * as utils from "../../utils";
 import { logBullet } from "../../utils";
@@ -27,8 +29,12 @@ const APPHOSTING_YAML_TEMPLATE = readTemplateSync("init/apphosting/apphosting.ya
 /**
  * Set up an apphosting.yaml file for a new App Hosting project.
  */
-export async function doSetup(setup: Setup, config: Config): Promise<void> {
+export async function doSetup(setup: Setup, config: Config, options: Options): Promise<void> {
+  // Use dynamicImport to bypass Node ESM linkage cycles during mocha test loading.
+  const { dynamicImport } = eval("require")("../../dynamicImport");
+
   const projectId = setup.projectId as string;
+
   if (!(await isBillingEnabled(setup))) {
     throw new FirebaseError(
       `Firebase App Hosting requires billing to be enabled on your project. To upgrade, visit the following URL: https://console.firebase.google.com/project/${projectId}/usage/details`,
@@ -85,6 +91,17 @@ export async function doSetup(setup: Setup, config: Config): Promise<void> {
       utils.logWarning(`Firebase web app not set`);
     }
 
+    const experiments = await dynamicImport("./experiments");
+    const prompts = await dynamicImport("./apphosting/prompts");
+
+    let runtime = experiments.isEnabled("abiu") ? prompts.DEFAULT_RUNTIME : undefined;
+    let automaticBaseImageUpdatesDisabled = experiments.isEnabled("abiu") ? false : undefined;
+
+    if (experiments.isEnabled("abiu") && !options.nonInteractive) {
+      runtime = await prompts.promptRuntime(projectId, location);
+      automaticBaseImageUpdatesDisabled = !(await prompts.promptAutomaticBaseImageUpdates());
+    }
+
     const createBackendSpinner = ora("Creating your new backend...").start();
     const backend = await createBackend(
       projectId,
@@ -93,7 +110,11 @@ export async function doSetup(setup: Setup, config: Config): Promise<void> {
       /* serviceAccount= */ null,
       /* repository= */ undefined,
       webApp?.id,
+      /* rootDir= */ "/",
+      runtime,
+      automaticBaseImageUpdatesDisabled,
     );
+
     createBackendSpinner.succeed(`Successfully created backend!\n\t${backend.name}\n`);
   }