+++ src/init/features/functions/index.ts
@@ -15,6 +15,7 @@ import {
 import { FirebaseError } from "../../../error";
 import { functionsOrigin, runtimeconfigOrigin } from "../../../api";
 import * as supported from "../../../deploy/functions/runtimes/supported";
+import * as experiments from "../../../experiments";
 
 const MAX_ATTEMPTS = 5;
 
@@ -177,6 +178,12 @@ async function languageSetup(setup: any): Promise<void> {
       value: "python",
     });
   }
+  if (experiments.isEnabled("functionsrunapionly")) {
+    choices.push({
+      name: "Dart",
+      value: "dart",
+    });
+  }
   const language = await select({
     message: "What language would you like to use to write Cloud Functions?",
     default: "javascript",
@@ -207,6 +214,18 @@ async function languageSetup(setup: any): Promise<void> {
       // In practical sense, latest supported runtime will not be a decomissioned runtime,
       // but in theory this doesn't have to be the case.
       cbconfig.runtime = supported.latest("python") as supported.ActiveRuntime;
+      break;
+    case "dart":
+      cbconfig.ignore = [
+        ".dart_tool",
+        ".git",
+        "firebase-debug.log",
+        "firebase-debug.*.log",
+        "*.local",
+      ];
+      // In practical sense, latest supported runtime will not be a decomissioned runtime,
+      // but in theory this doesn't have to be the case.
+      cbconfig.runtime = supported.latest("dart") as supported.ActiveRuntime;
       break;
   }
   setup.functions.languageChoice = language;