+++ src/mcp/onemcp/onemcp_server.ts
@@ -8,7 +8,7 @@ import {
   CallToolRequest,
 } from "@modelcontextprotocol/sdk/types.js";
 import { Client } from "../../apiv2";
-import { ServerTool } from "../tool";
+import { ServerTool, ServerToolMeta } from "../tool";
 import { McpContext, ServerFeature } from "../types";
 import { FirebaseError } from "../../error";
 import { ensure } from "../../ensureApiEnabled";
@@ -19,9 +19,15 @@ import { ensure } from "../../ensureApiEnabled";
 export class OneMcpServer {
   private listClient: Client;
   private callClient: Client;
+  /**
+   * @param feature The Firebase feature this server belongs to.
+   * @param serverUrl The base URL of the remote MCP server.
+   * @param meta Metadata to be attached to every tool from this server.
+   */
   constructor(
     private readonly feature: ServerFeature,
     private readonly serverUrl: string,
+    private readonly meta: ServerToolMeta,
   ) {
     this.listClient = new Client({
       urlPrefix: this.serverUrl,
@@ -52,11 +58,14 @@ export class OneMcpServer {
         mcp: {
           ...mcpTool,
           name: `${this.feature}_${mcpTool.name}`,
-          _meta: {
-            requiresAuth: true,
-          },
+          _meta: { ...this.meta, feature: this.feature },
         },
-        fn: (args: any, ctx: McpContext) => this.callTool(mcpTool.name, args, ctx),
+        fn: (
+          args: {
+            [x: string]: unknown;
+          },
+          ctx: McpContext,
+        ) => this.callTool(mcpTool.name, args, ctx),
         isAvailable: () => Promise.resolve(true),
       }));
     } catch (error) {
@@ -69,7 +78,13 @@ export class OneMcpServer {
   /**
    * Proxies a tool call to the remote MCP server.
    */
-  private async callTool(toolName: string, args: any, ctx: McpContext): Promise<CallToolResult> {
+  private async callTool(
+    toolName: string,
+    args: {
+      [x: string]: unknown;
+    },
+    ctx: McpContext,
+  ): Promise<CallToolResult> {
     // TODO: Optimize this to not call ensure on every tool call.
     await ensure(ctx.projectId, this.serverUrl, this.feature, /* silent=*/ true);
     try {