+++ firebase-vscode/CHANGELOG.md
@@ -1,5 +1,25 @@
 ## NEXT
 
+## 2.2.3
+
+- Update internal `firebase-tools` dependency to 15.13.0
+
+## 2.2.2
+
+- Update internal `firebase-tools` dependency to 15.12.0
+
+## 2.2.1
+
+- Update internal `firebase-tools` dependency to 15.10.0
+
+## 2.2.0
+
+- Update internal `firebase-tools` dependency to 15.9.0
+
+## 2.1.2
+
+- Update internal `firebase-tools` dependency to 15.7.0
+
 ## 2.1.1
 
 - Update internal `firebase-tools` dependency to 15.5.1